+++ implementation/rocksdb/db/zns_impl/index/zns_version_edit.h
@@ -17,7 +17,7 @@
 #include "rocksdb/status.h"
 
 namespace ROCKSDB_NAMESPACE {
-typedef std::set<std::pair<int, uint64_t>> DeletedZoneSet;
+typedef std::set<std::pair<uint8_t, uint64_t>> DeletedZoneSet;
 /**
  * @brief Prepares the changes to the index structure to allow for CoW behaviour
  * for the index.
@@ -28,11 +28,11 @@ class ZnsVersionEdit {
   ~ZnsVersionEdit() = default;
 
   void Clear();
-  void AddSSDefinition(const size_t level, const uint64_t number,
+  void AddSSDefinition(const uint8_t level, const uint64_t number,
                        const uint64_t lba, const uint64_t lba_count,
                        const uint64_t numbers, const InternalKey& smallest,
                        const InternalKey& largest);
-  void RemoveSSDefinition(const size_t level, const uint64_t number);
+  void RemoveSSDefinition(const uint8_t level, const uint64_t number);
   // Used for Manifest logic
   void EncodeTo(std::string* dst) const;
   Status DecodeFrom(const Slice& src);
@@ -55,9 +55,9 @@ class ZnsVersionEdit {
   friend class ZnsVersionSet;
   friend class ZnsCompaction;
 
-  std::vector<std::pair<size_t, SSZoneMetaData>> new_ss_;
+  std::vector<std::pair<uint8_t, SSZoneMetaData>> new_ss_;
   DeletedZoneSet deleted_ss_;
-  std::vector<std::pair<size_t, SSZoneMetaData>> deleted_ss_seq_;
+  std::vector<std::pair<uint8_t, SSZoneMetaData>> deleted_ss_seq_;
   SequenceNumber last_sequence_;
   bool has_last_sequence_;
   std::string comparator_;
