+++ implementation/rocksdb/db/zns_impl/index/zns_version_set_builder.cc
@@ -10,13 +10,13 @@ ZnsVersionSet::Builder::Builder(ZnsVersionSet* vset, ZnsVersion* base)
   base_->Ref();
   BySmallestKey cmp;
   cmp.internal_comparator = &vset_->icmp_;
-  for (size_t level = 0; level < ZnsConfig::level_count; level++) {
+  for (uint8_t level = 0; level < ZnsConfig::level_count; level++) {
     levels_[level].added_ss = new ZoneSet(cmp);
   }
 }
 
 ZnsVersionSet::Builder::~Builder() {
-  for (size_t level = 0; level < ZnsConfig::level_count; level++) {
+  for (uint8_t level = 0; level < ZnsConfig::level_count; level++) {
     const ZoneSet* added = levels_[level].added_ss;
     std::vector<SSZoneMetaData*> to_unref;
     to_unref.reserve(added->size());
@@ -41,14 +41,14 @@ void ZnsVersionSet::Builder::Apply(const ZnsVersionEdit* edit) {
   // TODO Update compaction pointers
   // Delete files
   for (const auto& deleted_ss_set_kvp : edit->deleted_ss_) {
-    const int level = deleted_ss_set_kvp.first;
+    const uint8_t level = deleted_ss_set_kvp.first;
     const uint64_t number = deleted_ss_set_kvp.second;
     levels_[level].deleted_ss.insert(number);
   }
 
   // Add new files
   for (size_t i = 0; i < edit->new_ss_.size(); i++) {
-    const int level = edit->new_ss_[i].first;
+    const uint8_t level = edit->new_ss_[i].first;
     SSZoneMetaData* m = new SSZoneMetaData(edit->new_ss_[i].second);
     m->refs = 1;
     levels_[level].deleted_ss.erase(m->number);
@@ -60,7 +60,7 @@ void ZnsVersionSet::Builder::Apply(const ZnsVersionEdit* edit) {
 void ZnsVersionSet::Builder::SaveTo(ZnsVersion* v) {
   BySmallestKey cmp;
   cmp.internal_comparator = &vset_->icmp_;
-  for (size_t level = 0; level < ZnsConfig::level_count; level++) {
+  for (uint8_t level = 0; level < ZnsConfig::level_count; level++) {
     // Merge the set of added files with the set of pre-existing files.
     // Drop any deleted files.  Store the result in *v.
     const std::vector<SSZoneMetaData*>& base_ss = base_->ss_[level];
@@ -87,7 +87,7 @@ void ZnsVersionSet::Builder::SaveTo(ZnsVersion* v) {
 #ifndef NDEBUG
     // Make sure there is no overlap in levels > 0
     if (level > 0) {
-      for (uint32_t i = 1; i < v->ss_[level].size(); i++) {
+      for (size_t i = 1; i < v->ss_[level].size(); i++) {
         const InternalKey& prev_end = v->ss_[level][i - 1]->largest;
         const InternalKey& this_begin = v->ss_[level][i]->smallest;
         if (vset_->icmp_.Compare(prev_end, this_begin) >= 0) {
@@ -102,7 +102,7 @@ void ZnsVersionSet::Builder::SaveTo(ZnsVersion* v) {
   }
 }
 
-void ZnsVersionSet::Builder::MaybeAddZone(ZnsVersion* v, const size_t level,
+void ZnsVersionSet::Builder::MaybeAddZone(ZnsVersion* v, const uint8_t level,
                                           SSZoneMetaData* f) {
   if (levels_[level].deleted_ss.count(f->number) > 0) {
     // File is deleted: do nothing
