+++ implementation/rocksdb/db/zns_impl/table/zns_table_cache.cc
@@ -33,8 +33,8 @@ ZnsTableCache::ZnsTableCache(const Options& options,
 
 ZnsTableCache::~ZnsTableCache() { cache_.reset(); }
 
-Status ZnsTableCache::FindSSZone(const SSZoneMetaData& meta, const size_t level,
-                                 Cache::Handle** handle) {
+Status ZnsTableCache::FindSSZone(const SSZoneMetaData& meta,
+                                 const uint8_t level, Cache::Handle** handle) {
   Status s;
   char buf[sizeof(meta.number)];
   EncodeFixed64(buf, meta.number);
@@ -49,7 +49,7 @@ Status ZnsTableCache::FindSSZone(const SSZoneMetaData& meta, const size_t level,
 
 Iterator* ZnsTableCache::NewIterator(const ReadOptions& options,
                                      const SSZoneMetaData& meta,
-                                     const size_t level,
+                                     const uint8_t level,
                                      ZnsSSTable** tableptr) {
   if (tableptr != nullptr) {
     *tableptr = nullptr;
@@ -66,7 +66,7 @@ Iterator* ZnsTableCache::NewIterator(const ReadOptions& options,
 }
 
 Status ZnsTableCache::Get(const ReadOptions& options,
-                          const SSZoneMetaData& meta, const size_t level,
+                          const SSZoneMetaData& meta, const uint8_t level,
                           const Slice& key, std::string* value,
                           EntryStatus* status) {
   Cache::Handle* handle = nullptr;
